Reto 7: (Renamed)
- Tested-by: Cristian Rios (Creator)
- Status: Compleated
- Function: This is a calculator that has the next functions: addition, subtraction, multiplication, division, module, logarithm, power, square root.
- Other: This is a school proyect.
