--- CristianRR_POO/src/edu/Cristian_Rios/Reto7/process/RaizCuadrada.java
@@ -1,4 +1,0 @@
-package edu.Cristian_Rios.Reto7.process;
-
-public class RaizCuadrada {
-}
