+++ CristianRR_POO/src/edu/Cristian_Rios/Reto7/process/cal/Logaritmo.java
@@ -1,6 +1,6 @@
-package edu.Cristian_Rios.Reto7.process;
+package edu.Cristian_Rios.Reto7.process.cal;
 
-import edu.Cristian_Rios.Reto7.process.operacion_aritmetica;
+import edu.Cristian_Rios.Reto7.process.cal.Resta;
 
 public class Logaritmo extends Resta {
     public Logaritmo(float x, float y) {
