+++ CristianRR_POO/src/edu/Cristian_Rios/Reto7/process/cal/Multiplicacion.java
@@ -1,6 +1,6 @@
-package edu.Cristian_Rios.Reto7.process;
+package edu.Cristian_Rios.Reto7.process.cal;
 
-import edu.Cristian_Rios.Reto7.process.Suma;
+import edu.Cristian_Rios.Reto7.process.cal.Suma;
 
 public class Multiplicacion extends Suma {
     public Multiplicacion(float x, float y) {
