+++ CristianRR_POO/src/edu/Cristian_Rios/Reto7/process/cal/RaizCuadrada.java
@@ -1,0 +1,19 @@
+package edu.Cristian_Rios.Reto7.process.cal;
+
+import edu.Cristian_Rios.Reto7.process.operacion_aritmetica;
+
+public class RaizCuadrada extends Resta {
+    public RaizCuadrada(float x, float y) {
+        super(x, y);
+    }
+
+    @Override
+    public void calcular() {
+        // Utilizando el método de Newton-Raphson para calcular la raíz cuadrada
+        float estimacion = x / 2; // Estimación
+
+        // Añade aquí el código para calcular la raíz cuadrada
+
+    } // Aquí se espera el cierre del método calcular()
+
+} // Aquí se espera el cierre de la clase RaizCuadrada
