+++ CristianRR_POO/src/edu/Cristian_Rios/Reto7/process/cal/Suma.java
@@ -1,4 +1,4 @@
-package edu.Cristian_Rios.Reto7.process;
+package edu.Cristian_Rios.Reto7.process.cal;
 
 import edu.Cristian_Rios.Reto7.process.operacion_aritmetica;
